requirements.txt
twikit>=2.0.0
rookiepy>=0.5.0
pyyaml>=6.0
